Some more cleanup for Geonames importer, explicitly provide logback-test.xml in all projects.
Former-commit-id: 6f536da7e13300610af61cfc56e37f29a5c2baa4

+++ palladian-commons/src/main/java/ws/palladian/helper/math/FatStats.java
@@ -194,7 +194,7 @@ public class FatStats implements Stats {
         }
         List<Double> temp = getDoubleValues();
         Collections.sort(temp);
-        double n = (double)(p / 100.) * temp.size();
+        double n = p / 100. * temp.size();
         if (n == (int)n) {
             return 0.5 * temp.get((int)n - 1) + 0.5 * temp.get((int)n);
         } else {
